move techniquevm class
