+++ nav-app/src/app/matrix/technique-cell/contextmenu/contextmenu.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input, Output, EventEmitter, ElementRef, ViewEncapsulation } from '@angular/core';
-import { ContextMenuItem, Link } from '../../../classes';
+import { ContextMenuItem, Link, TechniqueVM } from '../../../classes';
 import { Technique, Tactic } from '../../../classes/stix';
-import { ViewModel, ViewModelsService, TechniqueVM } from '../../../services/viewmodels.service';
+import { ViewModel, ViewModelsService } from '../../../services/viewmodels.service';
 import { ConfigService } from '../../../services/config.service';
 import { CellPopover } from '../cell-popover';
